Fix: Fix video component using container presentation pattern
